Better User Matching / Searching
If there are multiple people with the same name, the app should not be able to display the right user (case sensitive), otherwise it will display that the user does not exist, e.g., Entering "john0" returns "john" and "John0" but displays neither has neither are exactly what the user searched for. Entering "John0" will return "John0", "John00", but only display "John0" since it is exactly who we searched for.